+++ src/mykrobe/base.py
@@ -136,6 +136,11 @@ genotyping_mixin.add_argument(
     default=False,
 )
 genotyping_mixin.add_argument(
+    "--dump_species_covgs",
+    help="Dump species probes coverage information to a JSON file",
+    metavar="FILENAME",
+)
+genotyping_mixin.add_argument(
     "-e",
     "--expected_error_rate",
     help="Expected sequencing error rate (default: %(default).3f)",
@@ -173,6 +178,12 @@ genotyping_mixin.add_argument(
     type=str,
     help="File path to save output file as. Default is to stdout",
     default="",
+)
+genotyping_mixin.add_argument(
+    "--ncbi_names",
+    help="Report NCBI species names in addiition to the usual species names in the JSON output. Only applies when the species is tb",
+    action="store_true",
+    default=False,
 )