+++ src/mykrobe/cmds/amr.py
@@ -167,20 +167,24 @@ def ref_data_from_args(args):
             "var_to_res_json": species_dir.json_file("amr"),
             "hierarchy_json": species_dir.json_file("hierarchy"),
             "lineage_json": species_dir.json_file("lineage"),
+            "ncbi_names_json": species_dir.json_file("ncbi_names"),
             "kmer": species_dir.kmer(),
             "version": species_dir.version(),
             "species_phylo_group": species_dir.species_phylo_group(),
         }
 
-    if ref_data["lineage_json"] is None:
-        ref_data["lineage_dict"] = None
-    else:
-        ref_data["lineage_dict"] = load_json(ref_data["lineage_json"])
+    for key in ["lineage", "ncbi_names"]:
+        if ref_data[f"{key}_json"] is None:
+            ref_data[f"{key}_dict"] = None
+        else:
+            ref_data[f"{key}_dict"] = load_json(ref_data[f"{key}_json"])
 
     return ref_data
 
 
-def detect_species_and_get_depths(cov_parser, hierarchy_json, wanted_phylo_group):
+def detect_species_and_get_depths(
+    cov_parser, hierarchy_json, wanted_phylo_group, probe_cov_json=None
+):
     depths = []
     if wanted_phylo_group is None:
         return {}, depths
@@ -193,6 +197,7 @@ def detect_species_and_get_depths(cov_parser, hierarchy_json, wanted_phylo_group
         species_covgs=cov_parser.covgs["species"],
         lineage_covgs=cov_parser.covgs.get("sub-species", {}),
         hierarchy_json_file=hierarchy_json,
+        probe_cov_json_file=probe_cov_json,
     )
     phylogenetics = species_predictor.run()
 
@@ -249,6 +254,13 @@ def fix_X_amino_acid_variants(sample_json):
         fix_amino_acid_X_variants_keys(sample_json["variant_calls"])
 
 
+def add_ncbi_species_names_to_phylo_dict(phylo, ncbi_names):
+    if "species" not in phylo or ncbi_names is None:
+        return
+    for species, species_d in phylo["species"].items():
+        species_d["ncbi_names"] = ncbi_names.get(species, "UNKNOWN")
+
+
 def run(parser, args):
     logger.info(f"Start runnning mykrobe predict. Command line: {' '.join(sys.argv)}")
     base_json = {args.sample: {}}
@@ -298,7 +310,10 @@ def run(parser, args):
         depths = [cp.estimate_depth()]
     else:
         phylogenetics, depths = detect_species_and_get_depths(
-            cp, ref_data["hierarchy_json"], ref_data["species_phylo_group"]
+            cp,
+            ref_data["hierarchy_json"],
+            ref_data["species_phylo_group"],
+            probe_cov_json=args.dump_species_covgs,
         )
 
     # Genotype
@@ -450,5 +465,9 @@ def run(parser, args):
 
     logger.info("Progress: writing output")
     fix_X_amino_acid_variants(base_json[args.sample])
+    if args.ncbi_names and ref_data["ncbi_names_dict"] is not None:
+        add_ncbi_species_names_to_phylo_dict(
+            base_json[args.sample]["phylogenetics"], ref_data["ncbi_names_dict"]
+        )
     write_outputs(args, base_json)
     logger.info("Progress: finished")